+++ litv-monitor-server/src/main/java/com/litv/monitor/dto/UserDTO.java
@@ -1,9 +1,11 @@
 package com.litv.monitor.dto;
 
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import jakarta.validation.constraints.NotBlank;
 import lombok.Data;
 
 @Data
+@JsonIgnoreProperties(ignoreUnknown = true)
 public class UserDTO {
 
     private Long id;